+++ LiftME/app/src/main/java/fnm/wrmc/nmmu/liftme/Trip.java
@@ -9,10 +9,10 @@ import java.text.SimpleDateFormat;
  */
 public class Trip implements Serializable {
 
-    private float pickupLat;
-    private float pickupLong;
-    private float destinationLat;
-    private float destinationLong;
+    private double pickupLat;
+    private double pickupLong;
+    private double destinationLat;
+    private double destinationLong;
     private Time pickupTime;
     private Time dropOffTime;
     private Date date;
@@ -45,35 +45,35 @@ public class Trip implements Serializable {
         this.date = date;
     }
 
-    public float getPickupLat() {
+    public double getPickupLat() {
         return pickupLat;
     }
 
-    public void setPickupLat(float pickupLat) {
+    public void setPickupLat(double pickupLat) {
         this.pickupLat = pickupLat;
     }
 
-    public float getPickupLong() {
+    public double getPickupLong() {
         return pickupLong;
     }
 
-    public void setPickupLong(float pickupLong) {
+    public void setPickupLong(double pickupLong) {
         this.pickupLong = pickupLong;
     }
 
-    public float getDestinationLat() {
+    public double getDestinationLat() {
         return destinationLat;
     }
 
-    public void setDestinationLat(float destinationLat) {
+    public void setDestinationLat(double destinationLat) {
         this.destinationLat = destinationLat;
     }
 
-    public float getDestinationLong() {
+    public double getDestinationLong() {
         return destinationLong;
     }
 
-    public void setDestinationLong(float destinationLong) {
+    public void setDestinationLong(double destinationLong) {
         this.destinationLong = destinationLong;
     }
